resolve all warnings
Bench: 22484598

+++ src/move_search/pvs.h
@@ -93,7 +93,7 @@ int q_search(Position &board, const int ply, int alpha, const int beta) {
 	Move best_move = Move();
 	MoveList<color> capture_moves(board, QSearchMoveGenerationsOptions);
 	ScoredMoves scored_moves = order_moves<color>(capture_moves, board, ply);
-	for (int move_idx = 0; move_idx < scored_moves.size(); move_idx++) {
+	for (int move_idx = 0; move_idx < static_cast<int>(scored_moves.size()); move_idx++) {
 		const Move legal_move = select_move(scored_moves, move_idx);
 		board.play<color>(legal_move);
 		data.q_nodes_searched += 1;
@@ -198,7 +198,7 @@ int pvs(Position &board, short depth, int ply, int alpha, int beta, bool do_null
 
 	Move best_move = select_move(scored_moves, 0);
 	int value = NEG_INF_CHESS;
-	for (int move_idx = 0; move_idx < scored_moves.size(); move_idx++) {
+	for (int move_idx = 0; move_idx < static_cast<int>(scored_moves.size()); move_idx++) {
 		Move legal_move = select_move(scored_moves, move_idx);
 
 		if (!pv_node && depth <= LMP_MIN_DEPTH && move_idx > depth * LMP_DEPTH_MULTIPLIER) {
